cyr_expire checks correct annotations with -p
When cyr_expire expunges a mailbox, it (by default) looks up an
annotation called "expire" on the mailbox and all it's ancestors (in
rootwards order) and uses the first found value to calculate how far
back in time the expunging should work.  However, if the -p option was
used to specify a prefix (i.e. root) for the mailbox tree traversal, the
annotation was being incorrectly fetched only from the prefix mailbox
and it's ancestors.