Handle CID clashing in APPEND, COPY, LMTP
When a message is added which would belong to two or more conversations,
the conversations are merged by choosing the maximum of the CIDs
involved as the winner and renaming all the losing CIDs to the winner in
the current transaction on the conversations db.  Mailbox actions are
also posted to cause CID renames in index files the next time the
mailbox is opened with LOCK_EXCLUSIVE.  To provide the data structures
to do those notifications, the cid -> folders mapping in the
conversations db is also updated on every message insert.