unit: allow more names for initialisation/cleanup
The suite initialisation and cleanup functions are detected by cunit.pl
using a naming convention.  The original convention was just to detect
the names "init" and "cleanup".  For consistentcy with other xUnit
implementations and with Kent Beck's original SmallTalk unit testing
paper, we now allow the names "setup" and "teardown", and variants
"set_up" and "tear_down".  For consistency with jUnit we also allow
"before" and "after".  All these names are now detected case
insensitively, so that "setUp" works, as does "SetUp" (and for that
matter "sEtuP").

Also, allow test functions to be named testFOO as well as test_FOO.  The
underscore after "test" is swallowed if present but is no longer
necessary.  This is for consistency with Perl's Test::Unit.