First cut at XCONVSORT.
Simple form of the syntax is identical to the SORT command.  Extended
form allows adding a (parenthesised) list of windowing arguments between
the sort arguments and the charset.  Currently the only windowing
argument is CONVERSATIONS which indicates that only the ID of the first
message to be encountered from each conversation is to be returned.