add highestmodseq to conversations db
This is actually a big change!  I had to move the struct
conversation_state into mailbox.c, as I have been saying
all along - so that mailbox_rewrite_index_record and
mailbox_append_index_record also do the right thing.  It's
actually a LOT nicer already like this, fewer call sites.
The only thing that would improve it further is having
the message parsing happen during the append/copy stage
as well, so that we could guarantee consistency without
having to deal with multiple locations.

Yes, you can argue about purity and that mailbox.c should
have a lower level thing that does nothing but the writes
to the file - but we already manage cache, quota, etc in
here.  It's the right place for this design, and redesigning
that is outside the scope of getting this working.