[IRIS-276] fix cleanup on bad cid in XCONVFETCH
Won't crash the imapd.

Also, passing the 0 cid results in a "BAD Invalid CID" response.
Passing a valid cid which is not present in the conversations
database, gives no results - not sure if this is good or not.

Also, when reporting the cid for the FETCH item, report the 0 cid
as NIL not 0000000000000000, as it used to do when first written.