annotate: do _fetch responses in a callback
Pull the code that emitted GETANNOTATION, GETMETADATA, and FETCH
ANNOTATION responses out into callbacks.  This reduces the complexity of
annotatemore_fetch() slightly; it now needs to pass less arguments
around and needs to know less about the context in which it's called.
More importantly however it allows using the _fetch code for other
things, like searching.