annotate: RFC5464 case-insensitive entry names
The RFC clearly states that entry names are case-insensitive, despite
the very similar entry names in RFC5257 being case sensitive.  So
enforce this.

Also, check the for /private and /shared at the start of entry names
more precisely.  Also, use the same and grammatically correct error
message for both locations where this is checked.