annotate: add annotation callout
A new imap option "annotation_callout" specifies a path to either an
executable to be forked and run or a UNIX domain socket to which a
service is listening.  During the APPEND path that program is fed
arguments including the new message filename, the body structure of the
message including file offsets, and any annotations and flags passed by
the client in the APPEND command.  The program's output is series of
data items similar to the arguments to the STORE command, and can be
used to set or clear flags and add or remove per-message annotations on
the message.  Program failures are logged and ignored.