Fix Read Uninit Memory in masterconf_getstring
In the case where the final parameter on the line is unquoted or is
quoted but incorrectly unterminated, masterconf_getstring() runs
blithely off the end of the line copying random characters from the heap
into a static buffer.

This has been broken since the current generation of the masterconf code
was written in Feb 2000.