Use storeargs.usinguid sensibly.
This flag is needed at various places inside index.c and is kept in
struct storeargs; this is all sensible.  However it's assigned inside
index_store() from a parameter passed down from cmd_store() which also
sets up the struct storeargs; this is not sensible.  Move the field
initialisation up to the caller and remove the parameter.