Invert STORE ANNOTATION control flow.
Previously, STORE ANNOTATION on multiple messages bypassed index_store()
and called annotatemore_store() with an annotate_scope_t set up to apply
to all the messages.  The loop over the sequence was done inside
annotatemore_store().

This worked but had the unfortunate side effect of bypassing all the
modseq logic in index_store, so that message modseqs were not updated
and the UNCHANGEDSINCE modifier was not enforced.

So invert the control flow.  Surgery required:

 - pass entryatts and other data needed for annotatemore down
   in struct storeargs
 - call index_store() even for STORE ANNOTATION
 - call annotatemore_store() from the message loop in index_store()
   via a helper
 - this is after the expunged flag is checked, so we can no longer
   incorrectly store annotations on expunged messages
 - this is also after the modseq flag is checked, so UNCHANGEDSINCE
   is now enforced for STORE ANNOTATION
 - the helper rewrites the mailbox record, ensuring the record modseq
   is updated
 - annotate_scope_t refers to a single message at a time
 - annotatemore_store() only stores for a single message at a time
 - annotatemore_store() doesn't eat it's arguments, so they can be
   used multiple times