annotate: per-mailbox annotation databases
Store per-message annotations in a separate annotations database for
each mailbox rather than in one global annotations database.  Per-server
and per-mailbox annotations are still kept in the global annotations
database.  This was considered necessary for performance and database
contention reasons.

The annotate module maintains an internal reference-counted list of
databases, keyed on mboxname (which is NULL for the global annotations
database).  Each database has it's own txn of course, but this is hidden
away inside the module.  The annotatemore_commit() and _abort() APIs
apply to all open databases which have outstanding changes (although we
should never actually see more than a single db with any writes, due to
the existing IMAP semantics).

The key format in the per-mailbox databases is different: the first part
is a UID instead of a mailbox name.

The global annotations db is opened at annotatemore_open() time, as
before.  All others the are opened and closed on demand, i.e. when
per-message annotations are stored, fetched, or iterated over.  At the
moment this leads to more than a sensible amount of database opening and
closing during some operations.  Per-mailbox annotations dbs are created
on demand, renamed when the mailbox is renamed, and unlinked with the
mailbox is deleted.

The imap option "metapartition_files" is extended to add the bit
"annotations" which controls whether the annotations db is placed on the
metadata partition.