master: centralise wait status handling
Put all the decoding of the exit status into a single function instead
of smearing it around several places.  This function works out whether a
process ended abnormally, emits the complaining syslogs if necessary,
and returns a boolean indicating whether the child failed.

Also, when defined use WCOREDUMP() to improve the syslog message when a
child dies due a signal to indicate whether a core was dumped.