annotate: don't use DB->rename or DB->remove
After talking with Bron and doing some reading, I see that all the other
code in the mailbox rename path merrily renames per-user databases using
the rename(2) system call on the filenames which are passed to
DB->open().  So if this ever becomes an inadequate procedure for
renaming databases, a lot more will be broken than just annotations.

Ditto with removing databases.

So stop using the DB->rename and DB->remove methods I recently added and
just call rename(2) and unlink(2).