fix sync when new message appears at both ends
If the sync master and replica discover that they have a pair of
messages with the same UID but different GUIDs, they go through a
complex dance to resolve the issue.  The end point of the dance is
supposed to see the two messages both stored safely at both ends, with
new UIDs and new MODSEQs.

However, copy_remote() on the master was failing to bump the modseq for
the message copied back from the replica when appending it to the
master's index.  So the actual end point of the dance had the following
problems:

 - master's copy of the new message from the replica has the replica's
   original modseq instead of a new one.

 - replica was missing the message it originally started with.

 - sync_client fails with a checksum error during full update.