Make quota -f less racy.
As discussed on the mailing list, make quota -f less racy against
updates to mailboxes that use more or less quota.

Broadly, the strategy is to leave the locking unchanged but use extra
state to track updates which lost the race against quota -f and account
for them at the end of the quota -f pass.

Two new pieces of state are used.

Firstly, each quota resource now has a second set of usage numbers,
called quota.usedBs[] due to lack of imagination.  They are stored in
the database as an optional third number in the tuple for each resource,
which is now (used, limit, usedB).  The usedBs are by default the
special value INVALID (which is non-zero) in memory to indicate that
they are not valid; INVALID numbers are not written out to the database
and numbers missing from the database become INVALID in-memory.

Secondly, each mailbox has a new flag SCANNED, saved in the index header
like all other flags.  This flag is normally unset.

The function quota_update_used(), which is called from mailbox_* and
mboxlist_* code to do the actual update to the quota db, now does one
extra thing: if the SCANNED flag was set on the mailbox and the usedB
value is not INVALID, usedB is incremented by the usage delta.

The quota -f algorithm now proceeds in three passes.

In pass1, we reset all the state to known values.  The SCANNED flag is
cleared on all mailboxes and their headers rewritten.  Then a
transaction is started on the quota db, quota records are read, their
usedB values are set to zero (presumably from INVALID, unless a previous
quota -f run crashed), and they are written out again.

In pass2, for each mailbox we add the mailbox' quota_mailbox_used field
to the running total, set the SCANNED flag on the mailbox, and write the
header out again.  This pass does not touch the quota db at all, but
works on in-memory copies of the quota information.

In pass3, we start a quota db transaction, and for each resource in each
quota root set it's usage to the newly calculated usage *plus* the
accumulated usedB value.  In the absence of racing updates to mailboxes,
the usedB values will be zero; otherwise they will account correctly for
any mailbox updates which occurred behind the quota -f scan, i.e. while
the quota -f scan was in progress but to mailboxes which had already
been scanned.  Finally, the usedB values are set back to INVALID, and
the quota db transaction is committed.

Note that at the end of the quota -f run, each mailbox has the SCANNED
flag set, but as this doesn't matter because the flag is only ever used
by quota -f and is reset in the pass1.

Also, added an undocumented -Z option to quota which is used for
reliably testing the race; it allows external test code like Cassandane
to start and stop the pass2 mailbox scan on a per mailbox basis.