unit: restore simplicity to quota TESTCASE macro
The whole point of this macro is to condense the code needed to run an
entire testcase down to a single line comprising a macro call with some
literals arguments for the inputs and the expected outputs, so that the
sequence of tests is nicely compact and visible.  This feature seems to
have been lost recently, restore it.