annotate: _delete and _msg_copy take mailbox*
so that the common code between them, _annotate_rewrite(), can have a
mailbox* for the old and new mailboxes.  This is the next step in the
plan to reimplement the annotate quota code to use mailbox* instead of
doing its own quota db transactions.

While we're changing the API, rename annotatemore_delete() to
annotate_delete() for consistency with other new API functions.

Some of the cunit test code now opens a real mailbox* every time it
needs one instead of faking one just for the .acl field.