annotate: use mailbox* instead of name in API
The annotate_state_t now always has a mailbox* open when used in
per-message or per-mailbox scope.  This required moving the traversal of
the mboxlist database to implement mailbox name patterns out of the API
and into the caller (although a helper annotate_apply_mailboxes() is
supplied to make it easier).

This has some implications.

Some of the internals of annotate_state_t needed reorganisation, which
impacted a lot of code.  Because a mailbox* is always open, the various
set/get functions which needed to do things to an open mailbox* are now
simpler.  Converesely however, we don't always have an mbentry and so
might need to do an extra mboxlist lookup (or more precisely, an
mboxlist lookup later and in a different place).

Because of the inversion of control flow, the operation of getting or
setting many annotations on many folders will now do more looking up of
entry names in the tables, which is now done once per matched folder
instead of once per SETMETADATA command.

The call to mboxlist_setspecialuse() is now double-opening a mailbox*
which fortunately the mailbox API can handle just fine, but it's
awkward.

The cunit annotate tests now all open a real mailbox* instead of faking
one.