annotate: track quota in mailbox header
using a previously spare 32b field to store the total usage in bytes of
all per-message and per-mailbox annotations attached to the mailbox.
Setting or deleting annotations adjusts this field in the existing open
mailbox* and dirties the quota information so that a quota db update
will be done in the caller's mailbox_close().  This should alleviate
several of the issues with the previous implementation of annotation
storage quotas.

Still to do: checking and accounting for annotation quota when adding
annotations in an APPEND.