Cleanups for annotate_getmailboxdb.
Remove the 'userannot' flag in the sync code.  It's superfluous; all it
does is give the same boolean result as the user_annot_db pointer, so
just test that instead.

Rename annotate_getmailboxdb to annotate_getdb and remove the
unnecessary @flags argument.  Rename the existing internal
annotate_getdb function as _annotate_getdb.  Handle the @mboxname
argument being NULL or "" correctly, so that the function never returns
a reference to the global db, as was the apparent intention.  Add a
comment in annotate.h documenting annotate_getdb().

Move struct annotate_db back into annotate.c.  It's not needed anywhere
outside there.