annotate: speed up FETCH 1:* ANNOTATION
by keeping an open reference on the per-message annotations db for the
selected mailbox if the FETCH command includes any ANNOTATIONs, thus
avoiding the slow open and close of the database for every single
message in the db.

It might be better to keep the annotations db reference in the
struct index_state, though.