Cleanup calls to annotatemore_rename()
In the usual case of not renaming a user, the RENAME code was passing a
`newuser' argument to annotatemore_rename() which was a pointer to an
uninitialised buffer on the stack.  This worked out ok, but only by
accident: it turns out that annotatemore_rename() doesn't use the
`newuser' argument if the `olduser' argument is NULL.

Now, the RENAME code doesn't rely on this accident anymore.