unit: fix bug with exit() catching
The exit interpolater is necessarily called on every exit().  It uses a
global state variable to decide whether it was called from code under
test (a bad thing) or the framework (a good thing) and complain mightily
in the former case.  That state variable was not being reset after tests
ran, in the case where a specific test was specified on the command line
and it happened to fail an ASSERT.