[IRIS-893] Fix COPY between users
The wrong conversation database was being opened in append_copy(), i.e.
the db for the source folder not the destination folder.  For copying
between any two folders of the same user, this doesn't matter, but it
breaks when copying between folders of different users.

This originally worked but regressed during one of the conversation API
rewrites.

Also, fix an incorrect comment for mboxname_parts_same_userid(); it
cannot return a negative error.