Use strarray_t properly in master
i.e. pass it around and store it as the actual data type in structures
instead of pretending we have only a char**.  This makes some code a lot
cleaner.  More importantly it allows us to free the tokenised command
lines stored in service->exec and event->exec correctly, thus fixing a
memleak.