master: tweak forkrate estimator
As we are now using microsecond resolution to measure the fork rate (as
opposed to the old second resolution code) we don't need to wait for a
full 2 second interval to elapse before getting our first forkrate
estimate.  This will have little effect in the field but allows the new
Cassandane forkrate test to converge faster.